VB-119: dispatch on client side
