+++ packages/frontend/components/filter/product-grid/component.tsx
@@ -9,14 +9,15 @@ export const ProductGrid = ({ lots }: ProductGridProps) => (
       {lots.map((item) => (
         <SplideSlide key={item.id}>
           <ProductCard
+            data={item}
             name={item.title}
             description={item.description}
             price={item.price}
             images={item.imageLinks}
-            currency="EURO"
+            currency="UAH"
             auctionDate={item.endDate}
             // eslint-disable-next-line @typescript-eslint/no-empty-function
-            onChangeIsFavorite={() => {}}
+            onButtonClick={() => {}}
           />
         </SplideSlide>
       ))}
